@@ -1,8 +1,11 @@
 source 'https://rubygems.org'
+ruby "2.2.0"
 
+gem 'rails', '~> 4.2.0'
 
-# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.1.4'
+gem "papertrail"
+gem "aws-sdk"
+
 # Use postgresql as the database for Active Record
 gem 'pg'
 # Use SCSS for stylesheets
@@ -49,3 +52,7 @@ gem 'omniauth'
 gem 'omniauth-facebook'
 gem 'validates_formatting_of'
 gem 'tzinfo-data'
+
+group :development do
+  gem "pry-rails"
+end
