@@ -1,32 +1,35 @@
 class AlbumsController < ApplicationController
-
-  #->Prelang (scaffolding:rails/scope_to_user)
   before_filter :require_user_signed_in, only: [:new, :edit, :create, :update, :destroy]
+  before_action :set_albums, only: [:index, :show]
+  before_action :set_album, only: [:edit, :update, :destroy]
 
-  before_action :set_album, only: [:show, :edit, :update, :destroy]
-
-  # GET /albums
-  # GET /albums.json
   def index
-    @albums = Album.all
+    @entries = [
+      "/test/downvote.gif",
+      "/test/did+you+fight+this+guy+by+any+chance+_33d2dec72fb5c4751e98273256381f0e.gif",
+      "/test/tumblr_lyh9zwsJTg1qifgls.gif",
+      "/test/sloth-follow-your-dreams.gif",
+      "/test/Popcorn-02-Stephen-Colbert.gif",
+      "/test/1948b.gif"
+    ]
   end
 
-  # GET /albums/1
-  # GET /albums/1.json
   def show
+    @entries = [
+      "/test/sloth-follow-your-dreams.gif",
+      "/test/Popcorn-02-Stephen-Colbert.gif",
+      "/test/1948b.gif"
+    ].shuffle
+    @album = Album.find_by_slug params[:id]
   end
 
-  # GET /albums/new
   def new
     @album = Album.new
   end
 
-  # GET /albums/1/edit
   def edit
   end
 
-  # POST /albums
-  # POST /albums.json
   def create
     @album = Album.new(album_params)
     @album.user = current_user
@@ -42,8 +45,6 @@ class AlbumsController < ApplicationController
     end
   end
 
-  # PATCH/PUT /albums/1
-  # PATCH/PUT /albums/1.json
   def update
     respond_to do |format|
       if @album.update(album_params)
@@ -56,8 +57,6 @@ class AlbumsController < ApplicationController
     end
   end
 
-  # DELETE /albums/1
-  # DELETE /albums/1.json
   def destroy
     @album.destroy
     respond_to do |format|
@@ -67,13 +66,17 @@ class AlbumsController < ApplicationController
   end
 
   private
-    # Use callbacks to share common setup or constraints between actions.
-    def set_album
-      @album = Album.find(params[:id])
-    end
 
-    # Never trust parameters from the scary internet, only allow the white list through.
-    def album_params
-      params.require(:album).permit(:name, :user_id)
-    end
+  def set_albums
+    @albums = Album.all
+  end
+
+  def set_album
+    @album = Album.find(params[:id])
+  end
+
+  # Never trust parameters from the scary internet, only allow the white list through.
+  def album_params
+    params.require(:album).permit(:name, :user_id)
+  end
 end
